requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.5
typing-extensions==4.13.2
